Add bitwise ops to `Integer` trait (#51)
Adds the following bitwise operations:

- `BitAnd`
- `BitOr`
- `BitXor`
- `Shl`
- `Shr`

Additionally adds impls of the `Bit*` traits on `UInt<LIMBS>` directly,
in addition to the existing impls on `Wrapping<UInt<LIMBS>>`.

This is safe because there is no difference between the "wrapping" or
"checked" versions of these operations: they always succeed and there is
no "wrapping" behavior to worry about.

This is debatably a breaking change but one that is extremely unlikely
to cause any problems in practice, especially since `crypto-bigint` v0.3
was just released. To be on the safe side though, these changes will
immediately be published in a v0.3.1 release, and the current v0.3.0
yanked so as to prevent any potential incompatibilities.