Replace `BoxedUint::new` with `::zero_with_precision` (#327)
The two APIs are nearly identical. The only difference is the latter is
infallible and panics if an invalid precision is given.

This also changes `BoxedUint::max` to have the same behavior.

Whether they should panic is a debatable point. Indeed having both
options would be good, but it should probably return `CtOption` where
`BoxedUint::max` was returning `Option`.

Regardless, this removes duplication and makes the API at least more
consistent with itself.