Inline low level multiplication and reduction functions (#776)
Fixes a performance regression introduced in #667. Evidently, compiler
relies a lot on knowing the slice sizes at compile time, so I'm inlining
`schoolbook_multiplication()`, `schoolbook_squaring()`, and
`montgomery_reduction_inner()`, so the compiler can optimize in case of
`Uint`s.

Benchmarks:
- `wrapping ops/split_mul, U256xU256` - 26ns to 9ns
- `Const Montgomery arithmetic/multiplication, U256*U256` - 41ns to 21ns
- `Dynamic Montgomery arithmetic/multiplication, U256*U256` - 62ns to
44ns

The effect is less pronounced for longer integers, but sill amounts to
5-10% speedup for U4096.

On a higher level, this affects many `crypto-primes` benchmarks, e.g.
doubles the speed of Lucas test for U128.

Possible addition: I think `panic!` in these functions can be replaced
with `debug_assert!`, but I don't insist on it.