Make `random_bits_core` platform independent (#781)
In its current state, `random_bits_core` behaves differently on 32-bit
and 64-bit targets in two ways:

1. The platforms order the randomly sampled bytes in a different order,
generating different outputs.
2. For bit_length $\mod 64 \in [1, 32]$, the 64-bit platform samples
four random bytes more than the 32-bit platform.
a. this means the state of the RNG after calling this function is
platform-dependent. This makes it impossible to use this function as an
rng-seed-to-integer map.

This patch introduces:
- benchmarks to gauge the performance of the `random_bits_core`
function,
- a patch to the `random_bits_core` function, and
- a test ensuring the `random_bits_core` output and rng state after
calling it are platform-independent.