Update safegcd GCD and modular inversion (#886)
This updates the safegcd implementation to remove the extra trait bounds
on the precomputed inverter and improve the performance. It implements
the safegcd-bounds half-delta step and bounds
(https://github.com/sipa/safegcd-bounds).

For the smallest sizes of Uint (U64 and maybe U128) it is currently
slightly slower, but the speed improves with larger sizes. For U256 is
it approximately 50% faster than the old implementation or bingcd.
`invert_odd_mod` is over 90% faster for U256.

Fixes #634

---------

Signed-off-by: Andrew Whitehead <cywolf@gmail.com>