Remove `PrecomputedInverter` and `Inverter` traits (#894)
`PrecomputedInverter` was largely a hack around the previous
implementation of safegcd, namely that it used a 62-bit unsaturated
implementation, which needed to be bounded by a different const generic
parameter `UNSAT_LIMBS`. Namely the `Invert` impl was bounded on the
presence of `PrecomputedInverter`, and the latter was only available for
certain limb sizes.

Now that safegcd no longer uses `UNSAT_LIMBS`, the traits and all of the
incidental complexity therein is unnecessary and can all be removed.

cc @andrewwhitehead